requirements.txt
boto3==1.40.53
googlesearch_python==1.3.0
httpx==0.28.1
python-dotenv==1.1.1
Crawl4AI==0.7.4
beautifulsoup4==4.12.2
pandas==2.1.3
pydantic==2.10.6

# Development
pytest==8.3.4

# Documentation
mkdocs==1.8.1
mkdocs-material==10.1.1
mkdocstrings==0.24.0

# Formatting
black==25.1.0

# Linting
flake8==7.0.0x